Improve test in contact.dao.test.

+++ de.hswt.hrm.test/src/de/hswt/hrm/test/database/AbstractDatabaseTest.java
@@ -1,14 +1,20 @@
 package de.hswt.hrm.test.database;
 
+import java.io.IOException;
+import java.nio.file.Paths;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Random;
 
+import static com.google.common.base.Strings.*;
+
 import org.junit.After;
 import org.junit.Before;
 
+import de.hswt.hrm.common.Config;
+import de.hswt.hrm.common.Config.Keys;
+import de.hswt.hrm.common.database.DatabaseFactory;
 import de.hswt.hrm.common.database.DatabaseUtil;
 import de.hswt.hrm.common.database.exception.DatabaseException;
 
@@ -20,36 +26,8 @@ import de.hswt.hrm.common.database.exception.DatabaseException;
  * You can use {@link #resetDatabase()} to reset the database during a test.
  */
 public abstract class AbstractDatabaseTest {
-
 	private Connection con;
 	private String dbName;
-	
-	public Connection getConnection() throws DatabaseException {
-        // load mariadb driver
-        try {
-            Class.forName("org.mariadb.jdbc.Driver");
-        }
-        catch (ClassNotFoundException e) {
-            throw new DatabaseException("Database driver not found.", e);
-        }
-        
-        // TODO load connection string from configuration
-        String config = "jdbc:mysql://10.154.4.20";
-        String username = "root";
-        String password = "70b145pl4ch7";
-        
-        if (con == null) {
-	        try {
-	            con = DriverManager.getConnection(config, username, password);
-	        }
-	        catch (SQLException e) {
-	            // TODO maybe add specific information about the error
-	            throw new DatabaseException(e);
-	        }
-        }
-        
-        return con;
-    }
 	
 	private String getRandomString(int count) {
 		final String alphabet = "abcdefghijklmnopqrstuvwxyz";
@@ -64,9 +42,9 @@ public abstract class AbstractDatabaseTest {
 		return sb.toString();
 	}
 	
-	private String createUniqueName() throws DatabaseException {
-		String prefix = "hrmtest_";
-		Connection con = getConnection();
+	private String createUniqueName() throws DatabaseException, IOException {
+	    String prefix = "hrmtest_";
+		Connection con = DatabaseFactory.getConnection();
 		
 		String name = prefix + getRandomString(5);
 		while (DatabaseUtil.dbExists(con, name)) {
@@ -77,8 +55,16 @@ public abstract class AbstractDatabaseTest {
 	}
 	
     @Before
-    public void createDatabase() throws DatabaseException {
-        Connection con = getConnection();
+    public void createDatabase() throws DatabaseException, IOException {
+        // load configuration
+        Config config = Config.getInstance();
+        config.load(Paths.get("../resources/hrm.properties"));
+        
+        // remove database name
+        config.setProperty(Keys.DB_NAME, "");
+        
+        // Create database
+        con = DatabaseFactory.getConnection();
         String name = createUniqueName();
         
         DatabaseUtil.createDb(con, name);
@@ -88,6 +74,10 @@ public abstract class AbstractDatabaseTest {
 			Statement stmt = con.createStatement();
 			stmt.executeQuery("USE " + name);
 			dbName = name;
+			
+			// Configure new db name
+			config.setProperty(Keys.DB_NAME, name);
+			
 		}
 		catch (SQLException e) {
 			throw new DatabaseException(e);
@@ -96,6 +86,10 @@ public abstract class AbstractDatabaseTest {
     
     @After
     public void dropDatabase() throws DatabaseException {
+        if (isNullOrEmpty(dbName)) {
+            throw new IllegalStateException("Database name must not be null!");
+        }
+        
         try {
 			Statement stmt = con.createStatement();
 			stmt.executeQuery("DROP DATABASE " + dbName);
@@ -109,8 +103,9 @@ public abstract class AbstractDatabaseTest {
      * Reset the test database.
      * 
      * @throws DatabaseException If an error occurs during reset.
+     * @throws IOException 
      */
-    public void resetDatabase() throws DatabaseException {
+    public void resetDatabase() throws DatabaseException, IOException {
         dropDatabase();
         createDatabase();
     }
